% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_actual_month.R
\name{commision_actual_monthUI_right}
\alias{commision_actual_monthUI_right}
\title{生成区域}
\usage{
commision_actual_monthUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
commision_actual_monthUI_bottom()
}
