% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_management_month.R
\name{commision_management_monthUI}
\alias{commision_management_monthUI}
\title{返利生成生成器界面}
\usage{
commision_management_monthUI(
  tabTitle = "佣金月度汇总表管理口径",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = commision_management_monthUI_left,
  func_right = commision_management_monthUI_right,
  func_bottom = commision_management_monthUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
返利生成生成器界面
}
\examples{
commision_management_monthUI()
}
