% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_management_month.R
\name{commision_management_monthUI_bottom}
\alias{commision_management_monthUI_bottom}
\title{生成区域}
\usage{
commision_management_monthUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
commision_management_monthUI_bottom()
}
