% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_management_month.R
\name{commision_management_monthUI_left}
\alias{commision_management_monthUI_left}
\title{请输入文件}
\usage{
commision_management_monthUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
commision_management_monthUI_left()
}
