% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_management.R
\name{commision_managementUI_bottom}
\alias{commision_managementUI_bottom}
\title{生成区域}
\usage{
commision_managementUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
commision_managementUI_bottom()
}
