% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_management.R
\name{commision_managementUI_left}
\alias{commision_managementUI_left}
\title{请输入文件}
\usage{
commision_managementUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
commision_managementUI_left()
}
