% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grossProfit_management.R
\name{grossProfit_managementUI}
\alias{grossProfit_managementUI}
\title{返利生成生成器界面}
\usage{
grossProfit_managementUI(
  tabTitle = "集团毛利明细表管理口径 ",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = grossProfit_managementUI_left,
  func_right = grossProfit_managementUI_right,
  func_bottom = grossProfit_managementUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
返利生成生成器界面
}
\examples{
grossProfit_managementUI()
}
