% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grossProfit_management.R
\name{grossProfit_managementUI_left}
\alias{grossProfit_managementUI_left}
\title{请输入文件}
\usage{
grossProfit_managementUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
grossProfit_managementUI_left()
}
