% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grossProfit_management.R
\name{grossProfit_managementUI_right}
\alias{grossProfit_managementUI_right}
\title{生成区域}
\usage{
grossProfit_managementUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
grossProfit_managementUI_bottom()
}
