% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_actual_year.R
\name{rebate_actual_yearUI_bottom}
\alias{rebate_actual_yearUI_bottom}
\title{生成区域}
\usage{
rebate_actual_yearUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
rebate_actual_yearUI_bottom()
}
