% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_detailAmt_actual_month.R
\name{rebate_detailAmt_actual_monthUI_right}
\alias{rebate_detailAmt_actual_monthUI_right}
\title{生成区域}
\usage{
rebate_detailAmt_actual_monthUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
rebate_detailAmt_actual_monthUI_bottom()
}
