% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_detailAmt_actual.R
\name{rebate_detailAmt_actualUI_left}
\alias{rebate_detailAmt_actualUI_left}
\title{请输入文件}
\usage{
rebate_detailAmt_actualUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
rebate_detailAmt_actualUI_left()
}
