% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_exchangerate.R
\name{rebate_exchangerateUI_left}
\alias{rebate_exchangerateUI_left}
\title{请输入文件}
\usage{
rebate_exchangerateUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
rebate_exchangerateUI_left()
}
