% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_exchangerate.R
\name{rebate_exchangerateUI_right}
\alias{rebate_exchangerateUI_right}
\title{生成区域}
\usage{
rebate_exchangerateUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
rebate_exchangerateUI_bottom()
}
