% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_management_month.R
\name{rebate_management_monthUI_left}
\alias{rebate_management_monthUI_left}
\title{请输入文件}
\usage{
rebate_management_monthUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
rebate_management_monthUI_left()
}
