% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_management_month.R
\name{rebate_management_monthUI_right}
\alias{rebate_management_monthUI_right}
\title{生成区域}
\usage{
rebate_management_monthUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
rebate_management_monthUI_bottom()
}
