% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_commisionAmount.R
\name{rule_commisionAmountUI}
\alias{rule_commisionAmountUI}
\title{返利生成生成器界面}
\usage{
rule_commisionAmountUI(
  tabTitle = "佣金规则表",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(7, 5, 12),
  func_left = rule_commisionAmountUI_left,
  func_right = rule_commisionAmountUI_right,
  func_bottom = rule_commisionAmountUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
返利生成生成器界面
}
\examples{
rule_commisionAmountUI()
}
