% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_commisionAmount.R
\name{rule_commisionAmountUI_left}
\alias{rule_commisionAmountUI_left}
\title{请输入文件}
\usage{
rule_commisionAmountUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
rule_commisionAmountUI_left()
}
