% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_commisionAmount.R
\name{rule_commisionAmountUI_right}
\alias{rule_commisionAmountUI_right}
\title{生成区域}
\usage{
rule_commisionAmountUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
rule_commisionAmountUI_bottom()
}
