% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_payment_month.R
\name{rule_payment_monthUI}
\alias{rule_payment_monthUI}
\title{返利生成生成器界面}
\usage{
rule_payment_monthUI(
  tabTitle = "支付金额USD表",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = rule_payment_monthUI_left,
  func_right = rule_payment_monthUI_right,
  func_bottom = rule_payment_monthUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
返利生成生成器界面
}
\examples{
rule_payment_monthUI()
}
