% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_rebateFobamount.R
\name{rule_rebateFobamountUI_left}
\alias{rule_rebateFobamountUI_left}
\title{请输入文件}
\usage{
rule_rebateFobamountUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
rule_rebateFobamountUI_left()
}
