% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_rebateFobamount.R
\name{rule_rebateFobamountUI_right}
\alias{rule_rebateFobamountUI_right}
\title{生成区域}
\usage{
rule_rebateFobamountUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
rule_rebateFobamountUI_bottom()
}
