% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_outStock_mngr.R
\name{sal_outStock_mngrUI}
\alias{sal_outStock_mngrUI}
\title{返利生成生成器界面}
\usage{
sal_outStock_mngrUI(
  tabTitle = "销售出库管理口径数据源",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = sal_outStock_mngrUI_left,
  func_right = sal_outStock_mngrUI_right,
  func_bottom = sal_outStock_mngrUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
返利生成生成器界面
}
\examples{
sal_outStock_mngrUI()
}
