% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardCost_management.R
\name{standardCost_managementUI_right}
\alias{standardCost_managementUI_right}
\title{生成区域}
\usage{
standardCost_managementUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
standardCost_managementUI_bottom()
}
