% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportFee_management.R
\name{transportFee_managementUI_left}
\alias{transportFee_managementUI_left}
\title{请输入文件}
\usage{
transportFee_managementUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
transportFee_managementUI_left()
}
