% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportFee_management.R
\name{transportFee_managementUI_right}
\alias{transportFee_managementUI_right}
\title{生成区域}
\usage{
transportFee_managementUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
transportFee_managementUI_bottom()
}
